Changed base container to map

+++ Air/AirInfo.cpp
@@ -79,13 +79,13 @@ void AirInfo::MoldLevel()
 {
 	auto moldRH = -0.0015*pow(this->temperature, 3) +
 		0.1193*pow(this->temperature, 2) - 2.9878*this->temperature + 102.96;
-	this->riskLevel = (this->humidity - moldRH);
+	this->riskLevel = static_cast<int>(this->humidity - moldRH);
 }
 void AirInfo::AvgMoldLevel()
 {
 	auto moldRH = -0.0015*pow(this->avgTemperature, 3) +
 		0.1193*pow(this->avgTemperature, 2) - 2.9878*this->avgTemperature + 102.96;
-	this->riskLevel = (this->avgHumidity - moldRH);
+	this->riskLevel = static_cast<int>(this->avgHumidity - moldRH);
 }
